Prepare to MMOEngine Wrapper. Warning! Part of the project is not compiled.

+++ TornadoEngine/Source/Base/Headers/CallBackRegistrator.h
@@ -89,7 +89,7 @@ public:
   }
   int RegisteredCount()
   {
-    return mMapObjFunc.size();
+    return (int)mMapObjFunc.size();
   }
   void Clear()
   {
